+++ go-away/src/output/kotlin/unions.rs
@@ -1,12 +1,9 @@
 use crate::{
-    output::prelude::*,
+    output::{kotlin::kserializer::KSerializer, prelude::*},
     types::{self, UnionRepresentation},
 };
 
-use super::{
-    data_classes::{DataClass, NewTypeClass},
-    to_camel_case,
-};
+use super::{data_classes::NewTypeClass, to_camel_case};
 
 pub struct Union<'a> {
     name: &'a str,
@@ -73,313 +70,224 @@ impl fmt::Display for Union<'_> {
         );
         writeln!(f, "}}")?;
 
+        let mut serializer = KSerializer::new(name);
         match &self.representation {
-            UnionRepresentation::AdjacentlyTagged { tag, content } => todo!(),
-            UnionRepresentation::InternallyTagged { tag } => todo!(),
+            UnionRepresentation::AdjacentlyTagged { tag, content } => {
+                serializer.serialize_body(AdjacentlyTaggedSerialize {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                    content,
+                });
+                serializer.deserialize_body(AdjacentlyTaggedDeserialize {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                    content,
+                });
+                serializer.descriptor(AdjacentlyTaggedDescriptor {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                    content,
+                });
+            }
+            UnionRepresentation::InternallyTagged { tag } => {
+                serializer.serialize_body(InternallyTaggedSerialize {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                });
+                serializer.deserialize_body(InternallyTaggedDeserialize {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                });
+                serializer.descriptor(InternallyTaggedDescriptor {
+                    name: self.name,
+                    variants: &self.variants,
+                    tag,
+                });
+            }
             UnionRepresentation::ExternallyTagged => {
-                writeln!(f, "object {serializer_name} : KSerializer<{name}> {{\n")?;
-                let f2 = &mut indented(f);
-                writeln!(
-                    f2,
-                    r#"override val descriptor: SerialDescriptor = buildClassSerialDescriptor("{name}") {{"#
-                )?;
-                writeln_for!(
-                    indented(f2),
-                    Variant { name: variant_name, serde_name, .. } in &self.variants,
-                    r#"element<{name}.{variant_name}>("{serde_name}", isOptional = true)"#
-                );
-                writeln!(f2, "}};\n")?;
-                let f2 = &mut indented(f);
-                writeln!(
-                    f2,
-                    "override fun serialize(encoder: Encoder, value: {name}) {{",
-                )?;
-                let f3 = &mut indented(f2);
-                writeln!(f3, "val composite = encoder.beginStructure(descriptor)")?;
-                writeln!(f3, "when(value) {{")?;
-                writedoc_for!(
-                    indented(f3),
-                    (i, Variant { name: variant_name, .. }) in self.variants.iter().enumerate(),
-                    r#"
-                        is {name}.{variant_name} ->
-                          composite.encodeSerializableElement(descriptor, {i}, {name}.{variant_name}.serializer(), value as {name}.{variant_name})
-                    "#
-                );
-                writeln!(f3, "}}")?;
-                writeln!(f3, "composite.endStructure(descriptor)")?;
-                writeln!(f2, "}}\n")?;
-                writeln!(f2, "override fun deserialize(decoder: Decoder): {name} {{")?;
-                let f3 = &mut indented(f2);
-                writeln!(f3, "val composite = decoder.beginStructure(descriptor)")?;
-                writeln!(
-                    f3,
-                    "val rv = when (val index = composite.decodeElementIndex(descriptor)) {{"
-                )?;
-                writedoc_for!(
-                    indented(f3),
-                    (i, Variant { name: variant_name, .. }) in self.variants.iter().enumerate(),
-                    r#"
-                        {i} -> composite.decodeSerializableElement(descriptor, {i}, {name}.{variant_name}.serializer())
-                    "#
-                );
-                writeln!(indented(f3), r#"else -> error("Unexpected input")"#)?;
-                writeln!(f3, "}}")?;
-                writeln!(f3, "composite.endStructure(descriptor)")?;
-                writeln!(f3, "return rv")?;
-                writeln!(f2, "}}")?;
-                writeln!(f, "}}")?;
+                serializer.serialize_body(ExternallyTaggedSerialize {
+                    name,
+                    variants: &self.variants,
+                });
+                serializer.deserialize_body(ExternallyTaggedDeserialize {
+                    name,
+                    variants: &self.variants,
+                });
+                serializer.descriptor(ExternallyTaggedDescriptor {
+                    name,
+                    variants: &self.variants,
+                });
             }
             UnionRepresentation::Untagged => todo!(),
         }
-
-        // Note: We won't need a Codable impl _if_ we've got an externally tagged union that
-        // has named struct fields inside it (which isn't even possible rn, :sigh:)
-        //
-        // In this case we'd just need some CodableKeys impls for each variant.
-
-        /* TODO:
-               let mut codable = Codable::new(self.name);
-
-               match &self.representation {
-                   UnionRepresentation::AdjacentlyTagged { tag, content } => {
-                       codable.encodable(AdjacentlyTaggedEncodable {
-                           name: self.name,
-                           variants: &self.variants,
-                           tag,
-                           content,
-                       });
-                       codable.decodable(AdjacentlyTaggedDecodable {
-                           name: self.name,
-                           variants: &self.variants,
-                           tag,
-                           content,
-                       });
-                   }
-                   UnionRepresentation::InternallyTagged { tag } => {
-                       codable.encodable(InternallyTaggedEncodable {
-                           name: self.name,
-                           variants: &self.variants,
-                           tag,
-                       });
-                       codable.decodable(InternallyTaggedDecodable {
-                           name: self.name,
-                           variants: &self.variants,
-                           tag,
-                       });
-                   }
-                   UnionRepresentation::ExternallyTagged => {
-                       codable.encodable(ExternallyTaggedEncodable {
-                           name,
-                           variants: &self.variants,
-                       });
-                       codable.decodable(ExternallyTaggedDecodable {
-                           name,
-                           variants: &self.variants,
-                       });
-                   }
-                   UnionRepresentation::Untagged => todo!("Support untagged variants for kotlin"),
-               }
-               writeln!(f, "{codable}")?;
-        */
+        writeln!(f, "{serializer}")?;
         Ok(())
     }
 }
 
-struct ExternallyTaggedEncodable<'a> {
+struct ExternallyTaggedSerialize<'a> {
     name: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for ExternallyTaggedEncodable<'_> {
+impl fmt::Display for ExternallyTaggedSerialize<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let ExternallyTaggedEncodable { name, variants } = self;
-        let coding_keys = format!("{name}.CodingKeys");
-
-        writeln!(
-            f,
-            "var container = encoder.container(keyedBy: {coding_keys}.self)"
-        )?;
-        writeln!(f, "switch self {{")?;
+        let ExternallyTaggedSerialize { name, variants } = self;
+        writeln!(f, "val composite = encoder.beginStructure(descriptor)")?;
+        writeln!(f, "when(value) {{")?;
         writedoc_for!(
             indented(f),
-            Variant { name, ..  } in variants.iter(),
+            (i, Variant { name: variant_name, .. }) in variants.iter().enumerate(),
             r#"
-                        case .{name}(let data):
-                            return try container.encode(data, forKey: .{name})
-                    "#
+                is {name}.{variant_name} ->
+                    composite.encodeSerializableElement(descriptor, {i}, {name}.{variant_name}.serializer(), value as {name}.{variant_name})
+            "#
         );
-        write!(f, "}}")
+        writeln!(f, "}}")?;
+        writeln!(f, "composite.endStructure(descriptor)")
     }
 }
 
-struct ExternallyTaggedDecodable<'a> {
+struct ExternallyTaggedDeserialize<'a> {
     name: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for ExternallyTaggedDecodable<'_> {
+impl fmt::Display for ExternallyTaggedDeserialize<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let ExternallyTaggedDecodable { name, variants } = self;
-        let coding_keys = format!("{name}.CodingKeys");
-
+        let ExternallyTaggedDeserialize { name, variants } = self;
+        writeln!(f, "val composite = decoder.beginStructure(descriptor)")?;
         writeln!(
             f,
-            "let container = try decoder.container(keyedBy: {coding_keys}.self)"
+            "val rv = when (val index = composite.decodeElementIndex(descriptor)) {{"
         )?;
         writedoc_for!(
-            &mut *f,
-            Variant { name, ty, ..  } in variants.iter(),
+            indented(f),
+            (i, Variant { name: variant_name, .. }) in variants.iter().enumerate(),
             r#"
-                if (container.contains(.{name})) {{
-                    self = .{name}(try container.decode({ty}.self, forKey: .{name}))
-                    return
-                }}
-            "#
+                        {i} -> composite.decodeSerializableElement(descriptor, {i}, {name}.{variant_name}.serializer())
+                    "#
         );
-        writedoc!(
-            f,
-            r#"
-                throw NSError(
-                    domain: "",
-                    code: 400,
-                    userInfo: [ NSLocalizedDescriptionKey: "Unknown variant of {name}"]
-                )"#
-        )
+        writeln!(indented(f), r#"else -> error("Unexpected input")"#)?;
+        writeln!(f, "}}")?;
+        writeln!(f, "composite.endStructure(descriptor)")?;
+        writeln!(f, "return rv")
+    }
+}
+struct ExternallyTaggedDescriptor<'a> {
+    name: &'a str,
+    variants: &'a [Variant<'a>],
+}
+
+impl fmt::Display for ExternallyTaggedDescriptor<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let ExternallyTaggedDescriptor { name, variants } = self;
+
+        writeln!(f, r#"buildClassSerialDescriptor("{name}") {{"#)?;
+        writeln_for!(
+            indented(f),
+            Variant { name: variant_name, serde_name, .. } in *variants,
+            r#"element<{name}.{variant_name}>("{serde_name}", isOptional = true)"#
+        );
+        writeln!(f, "}};\n")
     }
 }
 
 #[allow(dead_code)]
-struct InternallyTaggedEncodable<'a> {
+struct InternallyTaggedSerialize<'a> {
     name: &'a str,
     tag: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for InternallyTaggedEncodable<'_> {
+impl fmt::Display for InternallyTaggedSerialize<'_> {
     fn fmt(&self, _f: &mut fmt::Formatter<'_>) -> fmt::Result {
         todo!("Finish internally tagged enum support")
     }
 }
 
-struct InternallyTaggedDecodable<'a> {
+#[allow(dead_code)]
+struct InternallyTaggedDeserialize<'a> {
     name: &'a str,
     tag: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for InternallyTaggedDecodable<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let InternallyTaggedDecodable {
-            name,
-            variants,
-            tag,
-        } = self;
-        let coding_keys = format!("{name}.CodingKeys");
-
-        writedoc!(
-            f,
-            r#"
-                enum TagCoding: String, CodingKey {{
-                    case tag = "{tag}"
-                }}
-                let keyContainer = try decoder.container(keyedBy: TagCoding.self)
-                let key = try keyContainer.decode({coding_keys}.self, forKey: .tag)
-                switch key {{
-            "#
-        )?;
-        writedoc_for!(
-            indented(f),
-            Variant { name, ty, ..  } in *variants,
-            r#"
-                case .{name}:
-                    // Not 100% sure this'll work but
-                    let container = try decoder.singleValueContainer()
-                    self = .{name}(try container.decode({ty}.self))
-            "#
-        );
-        writeln!(f, "}}")
+impl fmt::Display for InternallyTaggedDeserialize<'_> {
+    fn fmt(&self, _f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        todo!("Finish internally tagged enum support")
     }
 }
 
 #[allow(dead_code)]
-struct AdjacentlyTaggedEncodable<'a> {
+struct InternallyTaggedDescriptor<'a> {
     name: &'a str,
     tag: &'a str,
-    content: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for AdjacentlyTaggedEncodable<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let AdjacentlyTaggedEncodable {
-            name,
-            variants,
-            tag,
-            content,
-        } = self;
-        let coding_keys = format!("{name}.CodingKeys");
-
-        writedoc!(
-            f,
-            r#"
-                enum ContainerKeys: String, CodingKey {{
-                    case tag = "{tag}"
-                    case data = "{content}"
-                }}
-                var container = encoder.container(keyedBy: ContainerKeys.self)
-                switch self {{
-            "#
-        )?;
-        writedoc_for!(
-            indented(f),
-            Variant { name,  ..  } in *variants,
-            r#"
-                case .{name}(let data):
-                    try container.encode({coding_keys}.{name}, forKey: .tag)
-                    try container.encode(data, forKey: .data)
-            "#
-        );
-        writeln!(f, "}}")
+impl fmt::Display for InternallyTaggedDescriptor<'_> {
+    fn fmt(&self, _f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        todo!("Finish internally tagged enum support")
     }
 }
 
-struct AdjacentlyTaggedDecodable<'a> {
+#[allow(dead_code)]
+struct AdjacentlyTaggedSerialize<'a> {
     name: &'a str,
     tag: &'a str,
     content: &'a str,
     variants: &'a [Variant<'a>],
 }
 
-impl fmt::Display for AdjacentlyTaggedDecodable<'_> {
+impl fmt::Display for AdjacentlyTaggedSerialize<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let AdjacentlyTaggedDecodable {
-            name,
-            variants,
-            tag,
-            content,
-        } = self;
-        let coding_keys = format!("{name}.CodingKeys");
-
-        writedoc!(
-            f,
-            r#"
-                enum ContainerKeys: String, CodingKey {{
-                    case tag = "{tag}"
-                    case data = "{content}"
-                }}
-                let container = try decoder.container(keyedBy: ContainerKeys.self)
-                let key = try container.decode({coding_keys}.self, forKey: .tag)
-                switch key {{
-            "#
-        )?;
+        let AdjacentlyTaggedSerialize { name, variants, .. } = self;
+        writeln!(f, "val composite = encoder.beginStructure(descriptor)")?;
+        writeln!(f, "when(value) {{")?;
         writedoc_for!(
             indented(f),
-            Variant { name, ty, ..  } in *variants,
+            Variant { name: variant_name, .. } in variants.iter(),
             r#"
-                case .{name}:
-                    self = .{name}(try container.decode({ty}.self, forKey: .data))
+                is {name}.{variant_name} ->
+                    composite.encodeSerializableElement(descriptor, 0, String.serializer(), "{name}")
+                    composite.encodeSerializableElement(descriptor, 1, {name}.{variant_name}.serializer(), value as {name}.{variant_name})
             "#
         );
-        writeln!(f, "}}")
+        writeln!(f, "}}")?;
+        writeln!(f, "composite.endStructure(descriptor)")
+    }
+}
+
+#[allow(dead_code)]
+struct AdjacentlyTaggedDeserialize<'a> {
+    name: &'a str,
+    tag: &'a str,
+    content: &'a str,
+    variants: &'a [Variant<'a>],
+}
+
+impl fmt::Display for AdjacentlyTaggedDeserialize<'_> {
+    fn fmt(&self, _f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        todo!("impl adjacent tagging for kotlin")
+    }
+}
+
+#[allow(dead_code)]
+struct AdjacentlyTaggedDescriptor<'a> {
+    name: &'a str,
+    tag: &'a str,
+    content: &'a str,
+    variants: &'a [Variant<'a>],
+}
+
+impl fmt::Display for AdjacentlyTaggedDescriptor<'_> {
+    fn fmt(&self, _f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        todo!("impl adjacent tagging for kotlin")
     }
 }