+++ go-away/tests/kotlin_output.rs
@@ -51,6 +51,7 @@ struct Two {
     y: bool,
 }
 
+/* TODO:
 #[test]
 fn test_adjacently_tagged_tuple_enums() {
     let mut registry = TypeRegistry::new();
@@ -59,7 +60,7 @@ fn test_adjacently_tagged_tuple_enums() {
     assert_snapshot!(go_away::registry_to_output::<go_away::KotlinType>(
         &registry
     ));
-}
+} */
 
 /*
 // TODO: Add StructEnum support...